Moved away from command to client side remote functions
